Refactor execution price fetch to use updater
