+++ src/cow-react/modules/limitOrders/updaters/QuoteUpdater/hooks/useQuoteRequestParams.ts
@@ -2,30 +2,51 @@ import { useLimitOrdersTradeState } from '@cow/modules/limitOrders/hooks/useLimi
 import { useWeb3React } from '@web3-react/core'
 import { LegacyFeeQuoteParams as FeeQuoteParams } from '@cow/api/gnosisProtocol/legacy/types'
 import { OrderKind } from '@cowprotocol/contracts'
-import { Token } from '@uniswap/sdk-core'
-import { parseUnits } from '@ethersproject/units'
+import { parseUnits } from 'ethers/lib/utils'
 import { useMemo } from 'react'
+import { DEFAULT_DECIMALS } from 'custom/constants'
+import { useTypedValue } from '@cow/modules/limitOrders/hooks/useTypedValue'
+import { getAddress } from '@cow/modules/limitOrders/utils/getAddress'
+import { calculateValidTo } from 'hooks/useSwapCallback'
+import { useUserTransactionTTL } from '@src/state/user/hooks'
 
 export function useQuoteRequestParams(): FeeQuoteParams | null {
-  const { inputCurrency, outputCurrency, inputCurrencyAmount, recipient, deadline } = useLimitOrdersTradeState()
+  const { inputCurrency, outputCurrency, recipient, orderKind } = useLimitOrdersTradeState()
   const { chainId, account } = useWeb3React()
+  const { exactTypedValue } = useTypedValue()
+  const [deadline] = useUserTransactionTTL()
 
-  const isFullInput = !!(deadline && chainId && inputCurrencyAmount && inputCurrency && outputCurrency)
+  return useMemo(() => {
+    if (!inputCurrency || !outputCurrency || !exactTypedValue || !chainId || !deadline) {
+      return null
+    }
 
-  const feeQuoteParams: FeeQuoteParams | null = isFullInput
-    ? {
-        chainId,
-        validTo: Math.round(Date.now() / 1000 + 60 * deadline),
-        receiver: recipient || account,
-        kind: OrderKind.SELL,
-        sellToken: (inputCurrency as Token).address,
-        buyToken: (outputCurrency as Token).address,
-        amount: parseUnits(inputCurrencyAmount.toExact(), inputCurrency?.decimals).toString(),
-        fromDecimals: inputCurrency.decimals,
-        toDecimals: outputCurrency.decimals,
-      }
-    : null
+    const fromDecimals = inputCurrency?.decimals ?? DEFAULT_DECIMALS
+    const toDecimals = outputCurrency?.decimals ?? DEFAULT_DECIMALS
 
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  return useMemo(() => feeQuoteParams, [JSON.stringify(feeQuoteParams)])
+    const amount =
+      orderKind === OrderKind.SELL
+        ? parseUnits(exactTypedValue, fromDecimals).toString()
+        : parseUnits(exactTypedValue, toDecimals).toString()
+
+    const sellToken = getAddress(inputCurrency)
+    const buyToken = getAddress(outputCurrency)
+    const validTo = calculateValidTo(deadline)
+
+    if (!amount || !sellToken || !buyToken) {
+      return null
+    }
+
+    return {
+      chainId,
+      validTo,
+      receiver: recipient || account,
+      kind: orderKind,
+      sellToken,
+      buyToken,
+      amount,
+      fromDecimals,
+      toDecimals,
+    }
+  }, [account, chainId, deadline, exactTypedValue, inputCurrency, orderKind, outputCurrency, recipient])
 }
