+++ src/cow-react/modules/limitOrders/utils/adjustDecimals.ts
@@ -1,5 +1,6 @@
 import BigNumber from 'bignumber.js'
+import { DEFAULT_DECIMALS } from 'custom/constants'
 
 export const adjustDecimals = (price: number, decimals: number) => {
-  return new BigNumber(price).div(10 ** (18 - decimals))
+  return new BigNumber(price).div(10 ** (18 - (decimals || DEFAULT_DECIMALS)))
 }
